+++ src/test/java/chess/domain/controller/ColumnTest.java
@@ -1,8 +1,9 @@
-package chess.view;
+package chess.domain.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
+import chess.controller.Column;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
